# docs/openapi3.json
{
  "components": {
...
    "schemas": {
...
      "github_com_ialekseychuk_my-place_internal_dto.AssignMultipleServicesToStaffRequest": {
        "properties": {
          "service_ids": {
            "items": {
              "type": "string"
            },
            "type": "array"
          }
        },
        "required": [
          "service_ids"
        ],
        "type": "object"
      },
      "github_com_ialekseychuk_my-place_internal_dto.AssignServiceToStaffRequest": {
        "properties": {
          "service_id": {
            "type": "string"
          }
        },
        "required": [
          "service_id"
        ],
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.CreateServiceRequest": {
        "properties": {
          "category_id": {
            "type": "string"
          },
          "duration_min": {
            "minimum": 1,
            "type": "integer"
          },
          "location_id": {
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "price_cents": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "required": [
          "duration_min",
          "name",
          "price_cents"
        ],
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.ErrorResponse": {
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.ServiceResponse": {
        "properties": {
          "business_id": {
            "type": "string"
          },
          "category_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "duration_min": {
            "type": "integer"
          },
          "id": {
            "type": "string"
          },
          "location_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "price_cents": {
            "type": "integer"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.StaffServiceResponse": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "service_id": {
            "type": "string"
          },
          "service_name": {
            "type": "string"
          },
          "staff_id": {
            "type": "string"
          },
          "staff_name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "github_com_ialekseychuk_my-place_internal_dto.UpdateServiceRequest": {
        "properties": {
          "category_id": {
            "type": "string"
          },
          "duration_min": {
            "minimum": 1,
            "type": "integer"
          },
          "location_id": {
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 3,
            "type": "string"
          },
          "price_cents": {
            "minimum": 1,
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "bearerFormat": "JWT",
        "description": "Bearer token (JWT) for authentication. Format: Bearer \u003ctoken\u003e",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "This is a sample API with OpenAPI documentation.",
    "title": "MyPlace API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/businesses/{businessID}/services": {
      "get": {
        "description": "Get all services for a business",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get services by business",
        "tags": [
          "Service"
        ]
      },
      "post": {
        "description": "Creates a new service for business",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.CreateServiceRequest"
              }
            }
          },
          "description": "Service object",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Create a new service",
        "tags": [
          "Service"
        ]
      }
    },
    "/api/v1/businesses/{businessID}/services/{serviceID}": {
      "delete": {
        "description": "Delete a service by ID",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Service ID",
            "in": "path",
            "name": "serviceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Service not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Delete service",
        "tags": [
          "Service"
        ]
      },
      "get": {
        "description": "Get a specific service by ID",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Service ID",
            "in": "path",
            "name": "serviceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Service not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get service by ID",
        "tags": [
          "Service"
        ]
      },
      "put": {
        "description": "Update service information",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Service ID",
            "in": "path",
            "name": "serviceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.UpdateServiceRequest"
              }
            }
          },
          "description": "Service update data",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Service not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Update service",
        "tags": [
          "Service"
        ]
      }
    },
    "/api/v1/businesses/{businessID}/staff-services/{staffID}/services": {
      "get": {
        "description": "Gets all services assigned to a staff member",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Get staff services",
        "tags": [
          "StaffServices"
        ]
      },
      "post": {
        "description": "Assigns a service to a staff member",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.AssignServiceToStaffRequest"
              }
            }
          },
          "description": "Service assignment request",
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.StaffServiceResponse"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff or service not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Assign service to staff",
        "tags": [
          "StaffServices"
        ]
      },
      "put": {
        "description": "Replaces all services assigned to a staff member with new ones",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.AssignMultipleServicesToStaffRequest"
              }
            }
          },
          "description": "Services assignment request",
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ServiceResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Bad request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Staff not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object"
                }
              }
            },
            "description": "Validation errors"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Replace staff services",
        "tags": [
          "StaffServices"
        ]
      }
    },
    "/api/v1/businesses/{businessID}/staff-services/{staffID}/services/{serviceID}": {
      "delete": {
        "description": "Removes a service assignment from a staff member",
        "parameters": [
          {
            "description": "Business ID",
            "in": "path",
            "name": "businessID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Staff ID",
            "in": "path",
            "name": "staffID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Service ID",
            "in": "path",
            "name": "serviceID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "401": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Assignment not found"
          },
          "500": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/github_com_ialekseychuk_my-place_internal_dto.ErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Unassign service from staff",
        "tags": [
          "StaffServices"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/api"
    }
  ]
}